fix(tests): fix type conversion errors in multiple test files
- Fix reader_service_test.go: correct ObjectID/string conversions
- Fix books_api_test.go: use ObjectID for ReadingProgress fields
- Fix relation_api_test.go: use ObjectID for IdentifiedEntity.ID
- Fix bookstore_converter_test.go: use float64 for Price field

+++ api/v1/bookstore/bookstore_converter_test.go
@@ -288,7 +288,7 @@ func TestPriceConversion(t *testing.T) {
 				Title:            "测试",
 				Author:           "作者",
 				Status:           bookstore.BookStatusDraft,
-				Price:            int64(tc.expected),
+				Price:            float64(tc.expected),
 			}
 
 			bookDTO := ToBookDTO(bookModel)
